split tests into cpu/gpu
